Implements version 1 of python machup
This first version of the python machup implements a linear
lifting-line solver and the supporting lifting-line grid and
geometry classes. Enough basic functionality is implemented to
handle a simple airplane case with uniform properties across the
wings, dihedral, and control surfaces. All results have been tested
against the existing version of machup to verify the correctness.