Implements aerodynamic twist
Implements linear interpolation of airfoil properties across
wingsegment to allow for a different airfoil to be specified at the
root and at the tip. Since the fortran machup interpolates lift and
moment coefficient linearly accross the wing and not airfoil
properties, special interpolations had to be added to allow for
comparison with the fortran version of the code.